backend/pyproject.toml
[tool.poetry]
name = "poetry"
version = "0.1.0"
description = ""
authors = ["kirill <Richi.mosberg@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10"
fastapi = {extras = ["all"], version = "^0.110.1"}
sqlalchemy = "^2.0.29"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
pyjwt = "^2.8.0"
bcrypt = "^4.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
